Move `CollectionEditForm` to its own file
